+++ src/main/java/io/github/thebusybiscuit/slimefun4/implementation/operations/FuelOperation.java
@@ -1,4 +1,4 @@
-package io.github.thebusybiscuit.slimefun4.core.machines;
+package io.github.thebusybiscuit.slimefun4.implementation.operations;
 
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
@@ -6,6 +6,7 @@ import javax.annotation.Nullable;
 import org.apache.commons.lang.Validate;
 import org.bukkit.inventory.ItemStack;
 
+import io.github.thebusybiscuit.slimefun4.core.machines.MachineOperation;
 import me.mrCookieSlime.Slimefun.Objects.SlimefunItem.abstractItems.MachineFuel;
 
 public class FuelOperation implements MachineOperation {