+++ src/main/java/io/github/thebusybiscuit/slimefun4/implementation/operations/MiningOperation.java
@@ -1,9 +1,11 @@
-package io.github.thebusybiscuit.slimefun4.core.machines;
+package io.github.thebusybiscuit.slimefun4.implementation.operations;
 
 import javax.annotation.Nonnull;
 
 import org.apache.commons.lang.Validate;
 import org.bukkit.inventory.ItemStack;
+
+import io.github.thebusybiscuit.slimefun4.core.machines.MachineOperation;
 
 public class MiningOperation implements MachineOperation {